# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/addarticle": {
            "post": {
                "description": "title category只能为0 或1 content token 为必填",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "私有方法"
                ],
                "summary": "添加文章接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文章分类 0为技术类 1为生活类",
                        "name": "category",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文章内容 ",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"添加成功！\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/deletearticle": {
            "delete": {
                "description": "identification token为必填",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "私有方法"
                ],
                "summary": "删除文章接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章唯一标识",
                        "name": "identification",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ \"code\": 1, \"err\": \"删除文章不存在或者已经删除！\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "用户名 密码 为必填",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "私有方法"
                ],
                "summary": "登录接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"msg\": \"登陆成功\",\"token\": \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpbmRlbnRseSI6IjZhMmE0NjJjLWExMDctNDhlYS04MmU1LTc0ZTMwODMyN2U2ZiIsInVzZXJuYW1lIjoiYWRtaW4iLCJpc3MiOiJ0ZXN0IiwiZXhwIjoxNjc4Nzg2NTM1fQ.P4dJ_f2UGhKbpiIqHxTxghRKwKIlCpF2XjryHCSnKKk\"}",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/api/updatearticle": {
            "get": {
                "description": "identification content category title 为必填",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "私有方法"
                ],
                "summary": "更新文章接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章唯一标识",
                        "name": "identification",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文章内容",
                        "name": "content",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文章分类",
                        "name": "category",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文章标题",
                        "name": "title",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ \"code\": 200, \"msg\": \"更新成功！\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/addcommmits": {
            "post": {
                "description": "article_id content token  为必填",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "发布评论接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章唯一标识",
                        "name": "article_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "父级评论",
                        "name": "parent_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "评论内容",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ \"code\": 200, \"msg\": \"更新成功！\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/articlelist": {
            "get": {
                "description": "token 为必填",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "获取文章接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页数,不填默认为1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "每页大小，默认为20",
                        "name": "number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200,\t\"data\": {\"length\": 4, \"list\": [\t{\t\"id\": 1, \"identification\": \"1068dcc0-4fd3-461d-9f4b-0100879457eb\",\t\"content\": \"1\",\t\"status\": 0,\t\"title\": \"1\",\t\"create_time\": \"1686055918562468\",\t\"ip\": \"127.0.0.1\", \"update_time\": \"1\", \"category\": \"1\" },]},\"msg\": \"获取数据成功！\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/commentlist": {
            "get": {
                "description": "article_id  token  为必填",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "获取评论接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章唯一标识",
                        "name": "article_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ \"code\": 200, \"msg\": \"更新成功！\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "用户名 密码 为必填",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "登录接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"msg\": \"登陆成功\",\"token\": \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpbmRlbnRseSI6IjZhMmE0NjJjLWExMDctNDhlYS04MmU1LTc0ZTMwODMyN2U2ZiIsInVzZXJuYW1lIjoiYWRtaW4iLCJpc3MiOiJ0ZXN0IiwiZXhwIjoxNjc4Nzg2NTM1fQ.P4dJ_f2UGhKbpiIqHxTxghRKwKIlCpF2XjryHCSnKKk\"}",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "description": "token 为必填",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "退出登录接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ \"code\": 200, \"msg\": \"退出登录成功！\" }",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "用户名 密码 邮箱 为必填",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公共方法"
                ],
                "summary": "注册接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\t\"code\": 200,\t\"msg\": \"注册成功！\" \t}",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        }
    }
}